Automated_Azure_VM/requirements.txt
azure-identity>=1.12.0
azure-mgmt-compute>=29.1.0
azure-mgmt-network>=25.1.0
azure-mgmt-resource>=23.0.1
azure-cli-core>=2.50.0